+++ druid-shell/src/platform/gtk/window.rs
@@ -41,7 +41,7 @@ use super::util::assert_main_thread;
 use crate::common_util::IdleCallback;
 use crate::dialog::{FileDialogOptions, FileDialogType, FileInfo};
 use crate::keyboard;
-use crate::mouse::{Cursor, MouseButton, MouseEvent};
+use crate::mouse::{Cursor, MouseButton, MouseButtons, MouseEvent};
 use crate::window::{IdleToken, Text, TimerToken, WinHandler};
 use crate::Error;
 
@@ -285,17 +285,21 @@ impl WindowBuilder {
             Inhibit(false)
         }));
 
-        win_state.drawing_area.connect_button_press_event(clone!(handle => move |_widget, button| {
+        win_state.drawing_area.connect_button_press_event(clone!(handle => move |_widget, event| {
             if let Some(state) = handle.state.upgrade() {
                 if let Ok(mut handler) = state.handler.try_borrow_mut() {
-                    handler.mouse_down(
-                        &MouseEvent {
-                            pos: Point::from(button.get_position()),
-                            count: get_mouse_click_count(button.get_event_type()),
-                            mods: get_modifiers(button.get_state()),
-                            button: get_mouse_button(button.get_button()),
-                        },
-                    );
+                    if let Some(button) = get_mouse_button(event.get_button()) {
+                        let button_state = event.get_state();
+                        handler.mouse_down(
+                            &MouseEvent {
+                                pos: Point::from(event.get_position()),
+                                buttons: get_mouse_buttons_from_modifiers(button_state).with(button),
+                                mods: get_modifiers(button_state),
+                                count: get_mouse_click_count(event.get_event_type()),
+                                button,
+                            },
+                        );
+                    }
                 } else {
                     log::info!("GTK event was dropped because the handler was already borrowed");
                 }
@@ -304,17 +308,21 @@ impl WindowBuilder {
             Inhibit(true)
         }));
 
-        win_state.drawing_area.connect_button_release_event(clone!(handle => move |_widget, button| {
+        win_state.drawing_area.connect_button_release_event(clone!(handle => move |_widget, event| {
             if let Some(state) = handle.state.upgrade() {
                 if let Ok(mut handler) = state.handler.try_borrow_mut() {
-                    handler.mouse_up(
-                        &MouseEvent {
-                            pos: Point::from(button.get_position()),
-                            mods: get_modifiers(button.get_state()),
-                            count: 0,
-                            button: get_mouse_button(button.get_button()),
-                        },
-                    );
+                    if let Some(button) = get_mouse_button(event.get_button()) {
+                        let button_state = event.get_state();
+                        handler.mouse_up(
+                            &MouseEvent {
+                                pos: Point::from(event.get_position()),
+                                buttons: get_mouse_buttons_from_modifiers(button_state).without(button),
+                                mods: get_modifiers(button_state),
+                                count: 0,
+                                button,
+                            },
+                        );
+                    }
                 } else {
                     log::info!("GTK event was dropped because the handler was already borrowed");
                 }
@@ -325,13 +333,13 @@ impl WindowBuilder {
 
         win_state.drawing_area.connect_motion_notify_event(clone!(handle => move |_widget, motion| {
             if let Some(state) = handle.state.upgrade() {
-
-                let pos = Point::from(motion.get_position());
+                let motion_state = motion.get_state();
                 let mouse_event = MouseEvent {
-                    pos,
-                    mods: get_modifiers(motion.get_state()),
+                    pos: Point::from(motion.get_position()),
+                    buttons: get_mouse_buttons_from_modifiers(motion_state),
+                    mods: get_modifiers(motion_state),
                     count: 0,
-                    button: get_mouse_button_from_modifiers(motion.get_state()),
+                    button: MouseButton::None,
                 };
 
                 if let Ok(mut handler) = state.handler.try_borrow_mut() {
@@ -346,13 +354,13 @@ impl WindowBuilder {
 
         win_state.drawing_area.connect_leave_notify_event(clone!(handle => move |_widget, crossing| {
             if let Some(state) = handle.state.upgrade() {
-
-                let pos = Point::from(crossing.get_position());
+                let crossing_state = crossing.get_state();
                 let mouse_event = MouseEvent {
-                    pos,
-                    mods: get_modifiers(crossing.get_state()),
+                    pos: Point::from(crossing.get_position()),
+                    buttons: get_mouse_buttons_from_modifiers(crossing_state),
+                    mods: get_modifiers(crossing_state),
                     count: 0,
-                    button: get_mouse_button_from_modifiers(crossing.get_state()),
+                    button: MouseButton::None,
                 };
 
                 if let Ok(mut handler) = state.handler.try_borrow_mut() {
@@ -764,37 +772,51 @@ fn make_gdk_cursor(cursor: &Cursor, gdk_window: &gdk::Window) -> Option<gdk::Cur
     )
 }
 
-fn get_mouse_button(button: u32) -> MouseButton {
+fn get_mouse_button(button: u32) -> Option<MouseButton> {
     match button {
-        1 => MouseButton::Left,
-        2 => MouseButton::Middle,
-        3 => MouseButton::Right,
-        4 => MouseButton::X1,
-        5 => MouseButton::X2,
-        _ => MouseButton::Left,
+        1 => Some(MouseButton::Left),
+        2 => Some(MouseButton::Middle),
+        3 => Some(MouseButton::Right),
+        // GDK X backend interprets button press events for button 4-7 as scroll events
+        8 => Some(MouseButton::X1),
+        9 => Some(MouseButton::X2),
+        _ => None,
     }
 }
 
-fn get_mouse_button_from_modifiers(modifiers: gdk::ModifierType) -> MouseButton {
-    match modifiers {
-        modifiers if modifiers.contains(ModifierType::BUTTON1_MASK) => MouseButton::Left,
-        modifiers if modifiers.contains(ModifierType::BUTTON2_MASK) => MouseButton::Middle,
-        modifiers if modifiers.contains(ModifierType::BUTTON3_MASK) => MouseButton::Right,
-        modifiers if modifiers.contains(ModifierType::BUTTON4_MASK) => MouseButton::X1,
-        modifiers if modifiers.contains(ModifierType::BUTTON5_MASK) => MouseButton::X2,
-        _ => {
-            //FIXME: what about when no modifiers match?
-            MouseButton::Left
-        }
+fn get_mouse_buttons_from_modifiers(modifiers: gdk::ModifierType) -> MouseButtons {
+    let mut buttons = MouseButtons::new();
+    if modifiers.contains(ModifierType::BUTTON1_MASK) {
+        buttons.insert(MouseButton::Left);
     }
+    if modifiers.contains(ModifierType::BUTTON2_MASK) {
+        buttons.insert(MouseButton::Middle);
+    }
+    if modifiers.contains(ModifierType::BUTTON3_MASK) {
+        buttons.insert(MouseButton::Right);
+    }
+    // TODO: Determine X1/X2 state (do caching ourselves if needed)
+    //       Checking for BUTTON4_MASK/BUTTON5_MASK does not work with GDK X,
+    //       because those are wheel events instead.
+    if modifiers.contains(ModifierType::BUTTON4_MASK) {
+        buttons.insert(MouseButton::X1);
+    }
+    if modifiers.contains(ModifierType::BUTTON5_MASK) {
+        buttons.insert(MouseButton::X2);
+    }
+    buttons
 }
 
-fn get_mouse_click_count(event_type: gdk::EventType) -> u32 {
+fn get_mouse_click_count(event_type: gdk::EventType) -> u8 {
     match event_type {
         gdk::EventType::ButtonPress => 1,
         gdk::EventType::DoubleButtonPress => 2,
         gdk::EventType::TripleButtonPress => 3,
-        _ => 0,
+        gdk::EventType::ButtonRelease => 0,
+        _ => {
+            log::warn!("Unexpected mouse click event type: {:?}", event_type);
+            0
+        }
     }
 }